requirements.txt
streamlit
openai
python-dotenv
PyPDF2
langchain
langchain-openai
langchain-community
faiss-cpu
streamlit-mic-recorder
